requirements.txt
Flask==3.0.0
google-cloud-datastore==2.20.1
python-jose
six
requests
authlib
python-dotenv
